Add header and media section
Addition of the header section. It consists of a
title on the left and 3 links on the right using
flexbox.

Addition of the media section which consists of
one container wrapping two other containers. The
first one being a title, a description, and a but-
ton. The second one being a picture.

Margin and padding, font type, and background col-
or have been made consistant.